--- apps/web/app/future/(shared-page-wrapper)/(no-layout)/settings/teams/[id]/profile/page.tsx
@@ -1,10 +1,0 @@
-import Page from "@pages/settings/teams/[id]/profile";
-import { _generateMetadata } from "app/_utils";
-
-export const generateMetadata = async () =>
-  await _generateMetadata(
-    (t) => t("profile"),
-    (t) => t("profile_team_description")
-  );
-
-export default Page;
