--- apps/web/app/future/(shared-page-wrapper)/(no-layout)/settings/teams/[id]/sso/page.tsx
@@ -1,10 +1,0 @@
-import Page from "@pages/settings/teams/[id]/sso";
-import { _generateMetadata } from "app/_utils";
-
-export const generateMetadata = async () =>
-  await _generateMetadata(
-    (t) => t("sso_configuration"),
-    (t) => t("sso_configuration_description")
-  );
-
-export default Page;
