+++ apps/web/app/future/(shared-page-wrapper)/(settings-layout)/settings/teams/[id]/appearance/page.tsx
@@ -1,5 +1,6 @@
-import Page from "@pages/settings/teams/[id]/appearance";
 import { _generateMetadata } from "app/_utils";
+
+import Page from "@calcom/features/ee/teams/pages/team-appearance-view";
 
 export const generateMetadata = async () =>
   await _generateMetadata(
