+++ apps/web/app/future/(shared-page-wrapper)/(settings-layout)/settings/teams/[id]/billing/page.tsx
@@ -1,10 +1,10 @@
-import Page from "@pages/settings/teams/[id]/members";
+import Page from "@pages/settings/billing/index";
 import { _generateMetadata } from "app/_utils";
 
 export const generateMetadata = async () =>
   await _generateMetadata(
-    (t) => t("team_members"),
-    (t) => t("members_team_description")
+    (t) => t("billing"),
+    (t) => t("team_billing_description")
   );
 
 export default Page;
