+++ apps/web/app/future/(shared-page-wrapper)/(settings-layout)/settings/teams/[id]/members/page.tsx
@@ -1,10 +1,11 @@
-import Page from "@pages/settings/teams/[id]/billing";
 import { _generateMetadata } from "app/_utils";
+
+import Page from "@calcom/features/ee/teams/pages/team-members-view";
 
 export const generateMetadata = async () =>
   await _generateMetadata(
-    (t) => t("billing"),
-    (t) => t("team_billing_description")
+    (t) => t("team_members"),
+    (t) => t("members_team_description")
   );
 
 export default Page;
