+++ apps/web/app/future/(shared-page-wrapper)/(settings-layout)/settings/teams/[id]/sso/page.tsx
@@ -1,10 +1,11 @@
-import Page from "@pages/settings/teams/index";
 import { _generateMetadata } from "app/_utils";
+
+import Page from "@calcom/features/ee/sso/page/teams-sso-view";
 
 export const generateMetadata = async () =>
   await _generateMetadata(
-    (t) => t("teams"),
-    (t) => t("create_manage_teams_collaborative")
+    (t) => t("sso_configuration"),
+    (t) => t("sso_configuration_description")
   );
 
 export default Page;
